Extract salt state popup component
